Pt.2 Also extended type of EventLog::EventLogId to u64 since it is not that far-fetched that we can overflow u32 with the number of overall events in the DB. (#192)

+++ sidecar/src/sql/tables/deploy_event.rs
@@ -40,7 +40,7 @@ pub fn create_table_stmt() -> TableCreateStatement {
         .to_owned()
 }
 
-pub fn create_insert_stmt(event_log_id: u32, deploy_hash: String) -> SqResult<InsertStatement> {
+pub fn create_insert_stmt(event_log_id: u64, deploy_hash: String) -> SqResult<InsertStatement> {
     let insert_stmt = Query::insert()
         .into_table(DeployEvent::Table)
         .columns([DeployEvent::EventLogId, DeployEvent::DeployHash])
